+++ flowly4j-examplejpa/pom.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.flowly4j</groupId>
+        <artifactId>flowly4j</artifactId>
+        <version>0.25-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>flowly4j-examplejpa</artifactId>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.flowly4j</groupId>
+            <artifactId>flowly4j-core</artifactId>
+            <version>0.25-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.flowly4j</groupId>
+            <artifactId>flowly4j-jpa</artifactId>
+            <version>0.25-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
+
+        <!-- JDBC DataSources -->
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-c3p0</artifactId>
+            <version>6.2.6.Final</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-core</artifactId>
+            <version>6.2.7.Final</version>
+        </dependency>
+
+        <!-- MySql database -->
+        <dependency>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <version>8.0.28</version>
+        </dependency>
+        <!-- MySql database -->
+
+        <dependency>
+            <groupId>com.flowly4j</groupId>
+            <artifactId>flowly4j-mongodb</artifactId>
+            <version>0.25-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-deploy-plugin</artifactId>
+                <configuration>
+                    <skip>true</skip>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
