+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/CustomException.java
@@ -1,0 +1,16 @@
+package com.flowly4j.examplejpa;
+
+import com.flowly4j.core.tasks.compose.retry.stopping.Retryable;
+
+public class CustomException extends RuntimeException implements Retryable {
+
+    public CustomException(String message) {
+        super(message);
+    }
+
+    @Override
+    public Boolean canBeRetried() {
+        return true;
+    }
+
+}
