+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/Demo.java
@@ -1,0 +1,13 @@
+package com.flowly4j.examplejpa;
+
+
+public class Demo {
+
+    public static void main( String[] args ) {
+
+
+
+    }
+
+
+}
