+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/FinishA.java
@@ -1,0 +1,9 @@
+package com.flowly4j.examplejpa;
+
+import com.flowly4j.core.tasks.FinishTask;
+
+class FinishA extends FinishTask {
+
+
+
+}
