+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/FinishB.java
@@ -1,0 +1,9 @@
+package com.flowly4j.examplejpa;
+
+import com.flowly4j.core.tasks.FinishTask;
+
+class FinishB extends FinishTask {
+
+
+
+}
