+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/Person.java
@@ -1,0 +1,13 @@
+package com.flowly4j.examplejpa;
+
+import lombok.Value;
+
+import java.time.Instant;
+
+@Value(staticConstructor = "of")
+public class Person {
+    String name;
+    Integer age;
+    Instant bla;
+}
+
