+++ flowly4j-examplejpa/src/main/java/com/flowly4j/examplejpa/WorkflowA.java
@@ -1,0 +1,14 @@
+package com.flowly4j.examplejpa;
+
+import com.flowly4j.core.Workflow;
+import com.flowly4j.core.context.ExecutionContext.ExecutionContextFactory;
+import com.flowly4j.core.repository.Repository;
+import io.vavr.collection.List;
+
+public class WorkflowA extends Workflow {
+
+    public WorkflowA(Repository repository, ExecutionContextFactory executionContextFactory) {
+        super(new ExecutionTaskA(), repository, executionContextFactory, List.of(new ConsoleEventListener()));
+    }
+
+}
