+++ flowly4j-examplejpa/src/main/resources/META-INF/persistence.xml
@@ -1,0 +1,22 @@
+<persistence xmlns="http://java.sun.com/xml/ns/persistence"
+             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
+             version="2.0">
+
+    <persistence-unit name="com.flowly4j.examplejpa">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <class>com.flowly4j.jpa.SessionWrapper</class>
+        <class>com.flowly4j.jpa.ExecutionWrapper</class>
+        <class>com.flowly4j.jpa.AttemptsWrapper</class>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+        <properties>
+            <property name="jakarta.persistence.jdbc.driver" value="com.mysql.jdbc.Driver"/>
+            <!-- Reemplazar las properties url, user y password con los del sistema-->
+            <property name="jakarta.persistence.jdbc.url" value="jdbc:mysql://localhost:3306/test"/>
+            <property name="jakarta.persistence.jdbc.user" value="root"/>
+            <property name="jakarta.persistence.jdbc.password" value="0507"/>
+
+            <property name="hibernate.hbm2ddl.auto" value = "create"/>
+    </properties>
+    </persistence-unit>
+</persistence>
