--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/CustomException.java
@@ -1,16 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-import com.flowly4j.core.tasks.compose.retry.stopping.Retryable;
-
-public class CustomException extends RuntimeException implements Retryable {
-
-    public CustomException(String message) {
-        super(message);
-    }
-
-    @Override
-    public Boolean canBeRetried() {
-        return true;
-    }
-
-}
