--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/Demo.java
@@ -1,13 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-
-public class Demo {
-
-    public static void main( String[] args ) {
-
-
-
-    }
-
-
-}
