--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/DisjunctionA.java
@@ -1,25 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-import com.flowly4j.core.input.Key;
-import com.flowly4j.core.tasks.DisjunctionTask;
-import io.vavr.collection.List;
-
-
-public class DisjunctionA extends DisjunctionTask {
-
-    @Override
-    protected Boolean isBlockOnNoCondition() {
-        return true;
-    }
-
-    @Override
-    protected List<Branch> branches() {
-        return Branch.of(c -> true, new BlockingA(), new FinishB());
-    }
-
-    @Override
-    protected List<Key> customAllowedKeys() {
-        return List.empty();
-    }
-
-}
