--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/FinishB.java
@@ -1,9 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-import com.flowly4j.core.tasks.FinishTask;
-
-class FinishB extends FinishTask {
-
-
-
-}
