--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/Person.java
@@ -1,13 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-import lombok.Value;
-
-import java.time.Instant;
-
-@Value(staticConstructor = "of")
-public class Person {
-    String name;
-    Integer age;
-    Instant bla;
-}
-
