--- flowly4j-examplemariadb/src/main/java/com/flowly4j/examplemariadb/WorkflowA.java
@@ -1,14 +1,0 @@
-package com.flowly4j.examplemariadb;
-
-import com.flowly4j.core.Workflow;
-import com.flowly4j.core.context.ExecutionContext.ExecutionContextFactory;
-import com.flowly4j.core.repository.Repository;
-import io.vavr.collection.List;
-
-public class WorkflowA extends Workflow {
-
-    public WorkflowA(Repository repository, ExecutionContextFactory executionContextFactory) {
-        super(new ExecutionTaskA(), repository, executionContextFactory, List.of(new ConsoleEventListener()));
-    }
-
-}
