--- flowly4j-examplemariadb/src/main/resources/META-INF/persistence.xml
@@ -1,22 +1,0 @@
-<persistence xmlns="http://java.sun.com/xml/ns/persistence"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
-             version="2.0">
-
-    <persistence-unit name="com.flowly4j.examplemariadb">
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <class>com.flowly4j.mariadb.SessionWrapper</class>
-        <class>com.flowly4j.mariadb.ExecutionWrapper</class>
-        <class>com.flowly4j.mariadb.AttemptsWrapper</class>
-        <exclude-unlisted-classes>true</exclude-unlisted-classes>
-        <properties>
-            <property name="jakarta.persistence.jdbc.driver" value="com.mysql.jdbc.Driver"/>
-            <!-- Reemplazar las properties url, user y password con los del sistema-->
-            <property name="jakarta.persistence.jdbc.url" value="jdbc:mysql://localhost:3306/test"/>
-            <property name="jakarta.persistence.jdbc.user" value="root"/>
-            <property name="jakarta.persistence.jdbc.password" value="password"/>
-
-            <property name="hibernate.hbm2ddl.auto" value = "create"/>
-    </properties>
-    </persistence-unit>
-</persistence>
