+++ flowly4j-jpa/src/main/java/com/flowly4j/jpa/Product.java
@@ -1,0 +1,23 @@
+package com.flowly4j.jpa;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import lombok.AllArgsConstructor;
+import lombok.NonNull;
+import lombok.Value;
+import lombok.experimental.FieldNameConstants;
+
+@Value
+@AllArgsConstructor(staticName = "of", onConstructor = @__(@JsonCreator))
+@FieldNameConstants
+public class Product {
+
+    @NonNull
+    String transactionId;
+
+    @NonNull
+    String type;
+
+    @NonNull
+    String id;
+
+}
