+++ pom.xml
@@ -28,9 +28,9 @@
     <modules>
         <module>flowly4j-core</module>
         <module>flowly4j-mongodb</module>
-        <module>flowly4j-mariadb</module>
+        <module>flowly4j-jpa</module>
         <module>flowly4j-example</module>
-        <module>flowly4j-examplemariadb</module>
+        <module>flowly4j-examplejpa</module>
     </modules>
 
     <dependencyManagement>
